# web/assets/openapi.yaml
openapi: 3.0.3
info:
  title: RenoP HTTP API
  version: "2.0"
  contact:
    url: https://github.com/404Setup/RenoP
  description: |
    Management and registry APIs for RenoP, a self-hosted artifact repository server
    supporting Maven, unstructured file storage, npm, Cargo (Rust sparse index), and Docker/OCI registries.

    Default listen address: `0.0.0.0:3000`.

    **Route layout:**
    - Management API: `/api/...`
    - Maven artifact storage: `/{repo}/{path}` (Maven 2 layout)
    - Unstructured file storage: `/{repo}/{path}`
    - Docker/OCI Distribution API: `/v2/...`
    - Cargo sparse registry: `/{repo}/config.json` and `/{repo}/api/v1/...`
    - npm registry: `/{repo}/{package}` and `/{repo}/-/...`
    - Javadoc preview: `/javadoc/{repo}/{path}`
    - Cargo-doc preview: `/cargodoc/{repo}/{path}`

    **Wire format:**
    Most management endpoints use `application/x-protobuf`. Schemas here describe the
    logical fields (snake_case) for documentation and client stub generation.
    See `proto/api/v1/api.proto` for canonical protobuf definitions.
    Docker/OCI, npm, and Cargo endpoints use JSON.

    **Error bodies:**
    Management errors are plain text (`Unauthorized`, `Forbidden`, `Not found`).
    Docker browser-management errors additionally expose a stable `X-Renop-Error-Code` header for localization.
    Docker Distribution API errors follow `{"errors":[{"code":"...","message":"...","detail":"..."}]}`.
...

  license:
    name: MPL-2.0
...
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: renop_session
      description: HttpOnly session cookie issued by `POST /api/auth/login`. Idle timeout is approximately 7 days.
    basicAuth:
      type: http
      scheme: basic
      description: |
        `username:password` or `username:API-token` encoded as HTTP Basic credentials.
        Basic credentials are accepted only by standard package protocols, Docker token exchange, and chunked upload.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A prefixed API-token secret. API endpoints intersect the token's scopes with the owning account's current
        permissions. Query-string credentials and `Authorization: Session` are rejected.
  parameters:
...
    APITokenId:
      name: token_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Non-secret API-token identifier returned by the token metadata endpoints.
...
  responses:
    PlainUnauthorized:
      description: Not authenticated or credentials are invalid.
      content:
        text/plain:
          schema:
            type: string
            example: Unauthorized
    PlainForbidden:
      description: Authenticated but not permitted, or the account is expired.
      content:
        text/plain:
          schema:
            type: string
            example: Forbidden
    PlainNotFound:
      description: Resource not found. Some endpoints also return this for permission denials to avoid information disclosure.
      content:
        text/plain:
          schema:
            type: string
            example: Not found
    PlainBadRequest:
      description: Request parameters or body are invalid.
      content:
        text/plain:
          schema:
            type: string
            example: Bad Request
    PlainConflict:
      description: Conflict — resource already exists, an installation is already running, or a concurrent operation prevents this request.
      content:
        text/plain:
          schema:
            type: string
...
  schemas:
...
    APITokenScope:
      type: string
      enum:
        - repository:read
        - repository:publish
        - repository:delete
        - package:create
        - package:metadata
        - package:lifecycle
        - team:manage
        - domain:read
        - domain:create
        - domain:verify
        - domain:delete
        - package:manage
        - domain:manage
        - messages:read
        - account:read
        - account:write
        - statistics:read
        - admin:users
        - admin:repositories
        - admin:settings
        - admin:audit
        - admin:notifications
        - admin:updates
        - admin:statistics
    APIToken:
      type: object
      description: |
        Non-secret metadata for one durable fine-grained credential. `package:manage` and `domain:manage` may appear
        only on credentials created before capability splitting and cannot be assigned to new tokens.
      required: [ id, name, scopes, created_at ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 80
        scopes:
          type: array
          minItems: 1
          maxItems: 24
          uniqueItems: true
          items:
            $ref: "#/components/schemas/APITokenScope"
        targets:
          $ref: "#/components/schemas/APITokenTargets"
        created_at:
          type: integer
          format: int64
          description: Creation time as Unix milliseconds.
        expires_at:
          type: integer
          format: int64
          nullable: true
          description: Expiration time as Unix milliseconds; omitted or null means no expiration.
    APITokenListResponse:
      type: object
      required: [ tokens, limit ]
      properties:
        tokens:
          type: array
          items:
            $ref: "#/components/schemas/APIToken"
        limit:
          type: integer
          example: 50
    APITokenScopesResponse:
      type: object
      required: [ scopes ]
      properties:
        scopes:
          type: array
          description: Scopes currently assignable by this account. Administrator scopes are omitted for non-administrators.
          items:
            $ref: "#/components/schemas/APITokenScope"
        target_kinds:
          type: object
          description: Target kind accepted by each target-restrictable scope.
          additionalProperties:
            type: string
            enum: [ repository, package, team, domain ]
        target_limit:
          type: integer
          example: 128
    APITokenTargets:
      type: object
      description: |
        Optional exact-target restrictions keyed by a selected scope. Omitted scopes remain valid for every target
        authorized to the owning account. Repository values are repository names; package values use
        `repository/package`; team values use `package/repository/package` or `domain/domain-name`; domain values are
        canonical domain names. At most 128 target values are accepted across the object.
      additionalProperties:
        type: array
        minItems: 1
        uniqueItems: true
        items:
          type: string
          maxLength: 512
    CreateAPITokenRequest:
      type: object
      required: [ name, scopes ]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        scopes:
          type: array
          minItems: 1
          maxItems: 24
          uniqueItems: true
          items:
            $ref: "#/components/schemas/APITokenScope"
        targets:
          $ref: "#/components/schemas/APITokenTargets"
        expires_at:
          type: integer
          format: int64
          nullable: true
          description: Unix milliseconds, from five minutes through five years after creation; null means no expiration.
    CreateAPITokenResponse:
      type: object
      description: The secret is returned only by the successful creation response and is never recoverable later.
      required: [ token, secret ]
      properties:
        token:
          $ref: "#/components/schemas/APIToken"
        secret:
          type: string
          minLength: 51
          maxLength: 51
          pattern: '^rnp_pat_[A-Za-z0-9_-]{43}$'
...
    SuperTeam:
      type: object
      required: [ prefix, name, description, created_by, role_level, member_count, created_at, updated_at ]
      properties:
        prefix:
          type: string
          minLength: 2
          maxLength: 64
          pattern: "^[a-z0-9](?:[a-z0-9_-]*[a-z0-9])$"
          description: Immutable namespace shared by every supported package engine.
        name:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 512
        created_by:
          type: string
          description: Current creator username, or the retained creation-time name after account deletion.
        role_level:
          type: integer
          minimum: 0
          maximum: 4
          description: Current account role. Zero is returned only for a system administrator who is not a member.
        member_count:
          type: integer
          minimum: 1
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
    SuperTeamMember:
      type: object
      required: [ username, level, added_at ]
      properties:
        username:
          type: string
        level:
          type: integer
          minimum: 1
          maximum: 4
        added_at:
          type: integer
          format: int64
    SuperTeamDetails:
      type: object
      required: [ team, members, administrator ]
      properties:
        team:
          $ref: "#/components/schemas/SuperTeam"
        members:
          type: array
          items:
            $ref: "#/components/schemas/SuperTeamMember"
        administrator:
          type: boolean
...
    PublicationQuotaConfig:
      type: object
      required: [ file_limit, byte_limit, publication_limit, period ]
      properties:
        file_limit: { type: integer, format: int64, minimum: 1, maximum: 10000000, default: 600 }
        byte_limit: { type: integer, format: int64, minimum: 1, maximum: 1125899906842624, default: 41943040 }
        publication_limit: { type: integer, format: int64, minimum: 1, maximum: 1000000, default: 20 }
        period: { type: string, enum: [ day, week, month ], default: month }
    PublicationQuotaOverride:
      type: object
      description: Omit every field to restore global defaults. Individual limits may be zero to prevent publication.
      properties:
        file_limit: { type: integer, format: int64, minimum: 0, maximum: 10000000, nullable: true }
        byte_limit: { type: integer, format: int64, minimum: 0, maximum: 1125899906842624, nullable: true }
        publication_limit: { type: integer, format: int64, minimum: 0, maximum: 1000000, nullable: true }
        period: { type: string, enum: [ day, week, month ], nullable: true }
        unlimited:
          type: boolean
          nullable: true
          description: Manager-only no-quota override.
    PublicationQuotaStatus:
      type: object
      required: [ owner_type, owner_key, file_limit, byte_limit, publication_limit, period, period_start,
                  period_end, files_used, bytes_used, publications_used, files_reserved, bytes_reserved,
                  publications_reserved, inherited, unlimited ]
      properties:
        owner_type: { type: string, enum: [ user, super_team ] }
        owner_key: { type: string }
        file_limit: { type: integer, format: int64, minimum: 0 }
        byte_limit: { type: integer, format: int64, minimum: 0 }
        publication_limit: { type: integer, format: int64, minimum: 0 }
        period: { type: string, enum: [ day, week, month ] }
        period_start: { type: integer, format: int64 }
        period_end: { type: integer, format: int64 }
        files_used: { type: integer, format: int64, minimum: 0 }
        bytes_used: { type: integer, format: int64, minimum: 0 }
        publications_used: { type: integer, format: int64, minimum: 0 }
        files_reserved: { type: integer, format: int64, minimum: 0 }
        bytes_reserved: { type: integer, format: int64, minimum: 0 }
        publications_reserved: { type: integer, format: int64, minimum: 0 }
        inherited: { type: boolean }
        unlimited: { type: boolean }
...
paths:
...
  /api/auth/profile/api-tokens:
    get:
      tags: [ auth ]
      summary: List own API-token metadata
      description: Returns bounded non-secret metadata. Plaintext credentials are never returned by list operations.
      operationId: getAuthProfileApiTokens
      security:
        - sessionCookie: [ ]
      responses:
        "200":
          description: Token metadata and the per-account limit.
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APITokenListResponse"
        "401":
          $ref: "#/components/responses/PlainUnauthorized"
    post:
      tags: [ auth ]
      summary: Create an API token
      description: |
        Creates a named credential with an optional expiration. Requested scopes must be assignable by the account;
        every request is also capped by the account's current permissions. The secret is shown exactly once.
      operationId: postAuthProfileApiToken
      security:
        - sessionCookie: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPITokenRequest"
      responses:
        "201":
          description: Token created. Store the one-time secret before discarding this response.
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAPITokenResponse"
        "400":
          $ref: "#/components/responses/PlainBadRequest"
        "401":
          $ref: "#/components/responses/PlainUnauthorized"
        "409":
          $ref: "#/components/responses/PlainConflict"
  /api/auth/profile/api-tokens/scopes:
    get:
      tags: [ auth ]
      summary: List scopes assignable by the current account
      operationId: getAuthProfileApiTokenScopes
      security:
        - sessionCookie: [ ]
      responses:
        "200":
          description: Scope catalog filtered by current account permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APITokenScopesResponse"
        "401":
          $ref: "#/components/responses/PlainUnauthorized"
  /api/auth/profile/api-tokens/{token_id}:
    delete:
      tags: [ auth ]
      summary: Revoke an API token
      description: Revocation takes effect immediately, including for cached authentication results.
      operationId: deleteAuthProfileApiToken
      security:
        - sessionCookie: [ ]
      parameters:
        - $ref: "#/components/parameters/APITokenId"
      responses:
        "204":
          description: Token revoked.
        "401":
          $ref: "#/components/responses/PlainUnauthorized"
        "404":
          $ref: "#/components/responses/PlainNotFound"
...
  /api/super-teams/{prefix}:
    parameters:
      - name: prefix
        in: path
        required: true
        schema: { type: string, minLength: 2, maxLength: 64 }
    get:
      tags: [ super-teams ]
      summary: Get a global team
      operationId: getSuperTeam
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "200":
          description: Team metadata and username-only member list.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuperTeamDetails" }
        "404": { $ref: "#/components/responses/PlainNotFound" }
    put:
      tags: [ super-teams ]
      summary: Update mutable global-team metadata
      description: T4 owners and system administrators may change the name and description; the prefix is immutable.
      operationId: updateSuperTeam
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ name, description ]
              properties:
                name: { type: string, maxLength: 80 }
                description: { type: string, maxLength: 512 }
      responses:
        "200":
          description: Updated team details.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuperTeamDetails" }
    delete:
      tags: [ super-teams ]
      summary: Delete a global team
      description: T4 owners and system administrators may delete a team only when it owns no resources and has no pending transfer. Pending invitations are cancelled atomically.
      operationId: deleteSuperTeam
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "204": { description: Team deleted. }
...
  /api/publication-quota:
    get:
      tags: [ publication-quotas ]
      summary: Get the current account publication quota
      operationId: getOwnPublicationQuota
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "200":
          description: Effective account quota and current-period usage.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicationQuotaStatus" }
        "401": { $ref: "#/components/responses/PlainUnauthorized" }
  /api/publication-quota/users/{username}:
    parameters:
      - { name: username, in: path, required: true, schema: { type: string } }
    get:
      tags: [ publication-quotas ]
      summary: Get an account publication quota
      description: The account itself or a system manager may read this status.
      operationId: getUserPublicationQuota
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "200":
          description: Effective quota and current-period usage.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicationQuotaStatus" }
        "403": { $ref: "#/components/responses/PlainForbidden" }
        "404": { $ref: "#/components/responses/PlainNotFound" }
    put:
      tags: [ publication-quotas ]
      summary: Override an account publication quota
      description: Manager-only. An empty object restores global defaults.
      operationId: putUserPublicationQuota
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicationQuotaOverride" }
      responses:
        "200":
          description: Updated effective quota.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicationQuotaStatus" }
        "400": { $ref: "#/components/responses/PlainBadRequest" }
        "403": { $ref: "#/components/responses/PlainForbidden" }
  /api/publication-quota/super-teams/{prefix}:
    parameters:
      - { name: prefix, in: path, required: true, schema: { type: string, minLength: 2, maxLength: 64 } }
    get:
      tags: [ publication-quotas, super-teams ]
      summary: Get a global-team publication quota
      description: Team members and system managers may read this status.
      operationId: getSuperTeamPublicationQuota
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "200":
          description: Effective quota and current-period usage.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicationQuotaStatus" }
        "403": { $ref: "#/components/responses/PlainForbidden" }
        "404": { $ref: "#/components/responses/PlainNotFound" }
    put:
      tags: [ publication-quotas, super-teams ]
      summary: Override a global-team publication quota
      description: Manager-only. An empty object restores global defaults.
      operationId: putSuperTeamPublicationQuota
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicationQuotaOverride" }
      responses:
        "200":
          description: Updated effective quota.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicationQuotaStatus" }
        "400": { $ref: "#/components/responses/PlainBadRequest" }
        "403": { $ref: "#/components/responses/PlainForbidden" }
  /api/settings/publication-quota:
    get:
      tags: [ settings, publication-quotas ]
      summary: Get global publication quota defaults
      operationId: getPublicationQuotaSettings
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "200":
          description: Global defaults.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicationQuotaConfig" }
        "403": { $ref: "#/components/responses/PlainForbidden" }
    put:
      tags: [ settings, publication-quotas ]
      summary: Update global publication quota defaults
      operationId: putPublicationQuotaSettings
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicationQuotaConfig" }
      responses:
        "200":
          description: Updated global defaults.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicationQuotaConfig" }
        "400": { $ref: "#/components/responses/PlainBadRequest" }
        "403": { $ref: "#/components/responses/PlainForbidden" }

  # ========== Independent reviews ==========
...